% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{inverse.cdf.rn}
\alias{inverse.cdf.rn}
\title{Generate random samples via inverse CDF method}
\usage{
inverse.cdf.rn(f_inv, n)
}
\arguments{
\item{f_inv}{A function that accepts a numeric vector of uniform(0,1) draws
and returns the corresponding inverse CDF as a numeric vector.}

\item{n}{Integer scalar; desired sample size.}
}
\value{
Numeric vector of length \code{n} containing the generated draws.
}
\description{
Given an inverse-CDF function \code{f_inv} this returns \code{n} iid draws
from the associated distribution by transforming Uniform(0,1).
}
\examples{
inv_exp <- function(u) { -log(1 - u) / 4 }
x <- inverse.cdf.rn(inv_exp, 100)
}
