% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-validation.R
\name{tune.k}
\alias{tune.k}
\title{Tune K for KNN using subset cross validation}
\usage{
tune.k(data, y, vars, trials, split, k.grid)
}
\arguments{
\item{data}{A data.frame with predictors and response.}

\item{y}{Character; response name.}

\item{vars}{Character vector; predictor names.}

\item{trials}{Integer; number of random splits per K.}

\item{split}{Numeric; training proportion.}

\item{k.grid}{Integer vector; values of K to evaluate.}
}
\value{
The K (from k.grid) that achieved highest mean accuracy.
}
\description{
Tune K for KNN using subset cross validation
}
\examples{
\dontrun{
tune.k(
  iris,
  y = "Species",
  vars = c("Sepal.Length", "Sepal.Width"),
  trials = 20,
  split = 0.7,
  k.grid = 1:10
)
}
}
